WINCE: Don't build angle by default
When having a directx sdk, angle does get auto enabled
for all windows platforms. Windows CE directx builds on top
of opengl es 2, so if there is directx, there is also a
opengl es 2. So don't add more layers of indirection for
Windows CE. As well angle does not build for Windows CE.

Change-Id: I211a1de29431cd05d5cd60188f503d2cd299c9b3
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>