QTreeView - allow users to control data in the treestructure
This patch allows to set which logical index the tree is in.

Before the tree always displayed data from the logical index 0,
but it is actually more likely that the user wants to have data
from visual index 0 (which can be done by special value -1).

There is nothing special about logical index 0, and not being
able to change the tree-data is just annoying.

Change-Id: Ib070ce93343a0d2fbac3ad5a42cb4359401ac87c
Reviewed-by: Stephen Kelly <stephen.kelly@kdab.com>