doc: Avoid duplicate widget example entries
The link to widget examples in QtCore was added
in 4519d810d1d2be3af17bdd77edac2361620f9d2d, which made
widget examples pop up twice in generated lists
(once for qtcore and once for qtwidgets documentation).

I'm not sure why it was originally added to that module, but
the snippets in the fridgemagnets example (which was the
example fixed in the commit) has been verified to still work.

Change-Id: I1b349f08349b5499929315d426224ba8638adc8e
Reviewed-by: Paul Olav Tvete <paul.tvete@digia.com>