Add directwrite support to the configure script
The configure.exe already supported the -directwrite flag for some time.
This commit adds this flag to the configure script as well so
environments which use MinGW as cross-compiler can now also
optionally enable DirectWrite support.

The original configure.exe pieces were added in commit
c0fed43b04dec8bd549043d3ea5e28908128082c of the Qt4 repository.
The directwrite argument which this commit adds to the configure
script and the behaviour of it should be exactly the same as the
current configure.exe.

To get Qt built with DirectWrite support using the MinGW-w64 toolchain
at least mingw-w64 trunk revision 5451 (Nov 10 2012) is needed

Change-Id: I519718068cf5732c7ebcbf0a0fafb6b747db0bcf
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>