OpenGL examples: Add error message about ANGLE on Windows
These OpenGL examples require a Desktop OpenGL and will
not build with OpenGL ES 2.0. This means those examples
do not build on Windows with the default configuration
using ANGLE.
ANGLE is wrapping OpenGL ES 2.0 to DirectX and does not support
the full Desktop OpenGL feature set.

Since this is confusing for Windows users that do not know about ANGLE
this patch adds an explicit error message describing the solution.
(configuring Qt with -opengl desktop)

Task-number: QTBUG-28590

Change-Id: I782e6830b9e282ddcc8a2ee0a47faf3579d36aab
Reviewed-by: hjk <qthjk@ovi.com>
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>