Remove the framework directories from qmake's default include dirs
On Mac, GCC and Clang print "(framework directory)" at the end of the
directory listing for includes if that's a framework dir. The directory
with "(framework directory)" at the end does not exist, so it doesn't
affect anything by being there.

But we don't have to keep the list longer just for that.

Change-Id: I3d031d3d15c75801ec0d6112b2c913bd63e5def3
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@digia.com>