Fix DnD when using QSimpleDrag.
qApp->topLevelAt() returns the QShapedPixmapWindow that the DnD
operation created, but what we want is the QWidgetWindow.

QWidgetWindow has the code to handle QDragMove events.

In Qt4, QShapedPixmapWidget had a parent, so was never returned
by qApp->topLevelWidgets().

In Qt5 we must filter it out. Bug is visible in the QNX plugin,
which is the only user of QSimpleDrag.

Change-Id: I920da86f3a1a92ce8e087f5948292fa4c68d4d81
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>