set pkg-config host_bins directly to install dir
qmake tries to replace all occurrences of the install prefix with a
literal ${prefix}. of course this would mess up build paths which happen
to live under the prefix (this is untypical, but still), which would
break the sed magic in the subsequent install step.
as qmake itself has no use for the .pc files, it's ok to target them
directly for the install dir. we do the same with the include and lib
dirs already.

Task-number: QTBUG-28807
Change-Id: I53d7f0f3b357f67f30bea07554794f60160376cd
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>