Don't assume m11/m22 represents the scale of a QTransform
We can still assume it when pulling out the scale of the glyph cache,
since we limit it to scaling, but when computing the scale from the
current matrix we have to decompose it.

We always use a positive scale in the glyph-cache, and let the drawing
code take care of any flipping of the coordinates.

Change-Id: Ie3c4f2d91008a9be8f89bef29c15d80b23fb8a82
Reviewed-by: Tor Arne Vestbø <tor.arne.vestbo@digia.com>