showIsFullscreen: only respect this hint for windows and dialogs
If the window or widget is a popup, ignore the hint. The
intention of the flag should be to show main windows etc in
fullscreen, and not all kinds of popups and tooltips. The user can always
call showFullscreen explicit when necessary.

This is a backport of 48c73540ad2a507963db0fba57484faf66462ad7.

We need it in stable to fix menus on BlackBerry platform, they're
appearing fullscreen.

Task-number: QTBUG-29969

Change-Id: Id0d6cfc194916aa7f993cde54b5a0002f60399e1
Reviewed-by: Thomas McGuire <thomas.mcguire@kdab.com>